Move messages implementations other than InternalFlagOnlyMessage from `.message` to `.message.data`
