+++ mirai-core/src/commonMain/kotlin/message/data/atImpl.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2021 Mamoe Technologies and contributors.
+ * Copyright 2019-2022 Mamoe Technologies and contributors.
  *
  * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
  * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
@@ -7,7 +7,7 @@
  * https://github.com/mamoe/mirai/blob/dev/LICENSE
  */
 
-package net.mamoe.mirai.internal.message
+package net.mamoe.mirai.internal.message.data
 
 import kotlinx.io.core.buildPacket
 import kotlinx.io.core.readBytes
@@ -64,17 +64,23 @@ internal fun At.toJceData(
 }
 
 
-internal val atAllData = ImMsgBody.Elem(
-    text = ImMsgBody.Text(
-        str = AtAll.display,
-        attr6Buf = buildPacket {
-            // MessageForText$AtTroopMemberInfo
-            writeShort(1) // const
-            writeShort(0) // startPos
-            writeShort(AtAll.display.length.toShort()) // textLen
-            writeByte(1) // flag, may=1
-            writeInt(0) // uin
-            writeShort(0) // const
-        }.readBytes()
+@Suppress("unused") // limit scope
+internal val AtAll.jceData
+    get() = atAllData
+
+private val atAllData by lazy {
+    ImMsgBody.Elem(
+        text = ImMsgBody.Text(
+            str = AtAll.display,
+            attr6Buf = buildPacket {
+                // MessageForText$AtTroopMemberInfo
+                writeShort(1) // const
+                writeShort(0) // startPos
+                writeShort(AtAll.display.length.toShort()) // textLen
+                writeByte(1) // flag, may=1
+                writeInt(0) // uin
+                writeShort(0) // const
+            }.readBytes()
+        )
     )
-)
+}
