+++ mirai-core/src/commonMain/kotlin/message/data/faceImpl.kt
@@ -1,13 +1,13 @@
 /*
- * Copyright 2019-2021 Mamoe Technologies and contributors.
+ * Copyright 2019-2022 Mamoe Technologies and contributors.
  *
- *  此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
- *  Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
+ * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
+ * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
  *
- *  https://github.com/mamoe/mirai/blob/master/LICENSE
+ * https://github.com/mamoe/mirai/blob/dev/LICENSE
  */
 
-package net.mamoe.mirai.internal.message
+package net.mamoe.mirai.internal.message.data
 
 import kotlinx.io.core.toByteArray
 import net.mamoe.mirai.internal.network.protocol.data.proto.HummerCommelem
@@ -17,7 +17,7 @@ import net.mamoe.mirai.message.data.Face
 import net.mamoe.mirai.utils.hexToBytes
 import net.mamoe.mirai.utils.toByteArray
 
-internal val FACE_BUF = "00 01 00 04 52 CC F5 D0".hexToBytes()
+private val FACE_BUF = "00 01 00 04 52 CC F5 D0".hexToBytes()
 
 internal fun Face.toJceData(): ImMsgBody.Face {
     return ImMsgBody.Face(
