+++ mirai-core/src/commonMain/kotlin/message/visitor/MessageVisitorEx.kt
@@ -10,9 +10,12 @@
 
 package net.mamoe.mirai.internal.message.visitor
 
-import net.mamoe.mirai.internal.message.*
+import net.mamoe.mirai.internal.message.data.ForwardMessageInternal
+import net.mamoe.mirai.internal.message.data.LongMessageInternal
+import net.mamoe.mirai.internal.message.data.MarketFaceImpl
+import net.mamoe.mirai.internal.message.data.MarketFaceInternal
 import net.mamoe.mirai.internal.message.flags.*
-import net.mamoe.mirai.internal.message.source.*
+import net.mamoe.mirai.internal.message.source.MessageSourceInternal
 import net.mamoe.mirai.message.data.MessageSource
 import net.mamoe.mirai.message.data.visitor.MessageVisitor
 import net.mamoe.mirai.utils.castOrNull
