+++ mirai-core/src/commonMain/kotlin/network/protocol/packet/chat/receive/MessageSvc.PbSendMsg.kt
@@ -20,12 +20,12 @@ import net.mamoe.mirai.contact.Stranger
 import net.mamoe.mirai.internal.QQAndroidBot
 import net.mamoe.mirai.internal.contact.groupCode
 import net.mamoe.mirai.internal.contact.uin
-import net.mamoe.mirai.internal.message.ForwardMessageInternal
+import net.mamoe.mirai.internal.message.data.ForwardMessageInternal
+import net.mamoe.mirai.internal.message.data.toPtt
 import net.mamoe.mirai.internal.message.source.OnlineMessageSourceToFriendImpl
 import net.mamoe.mirai.internal.message.source.OnlineMessageSourceToGroupImpl
 import net.mamoe.mirai.internal.message.source.OnlineMessageSourceToStrangerImpl
 import net.mamoe.mirai.internal.message.source.OnlineMessageSourceToTempImpl
-import net.mamoe.mirai.internal.message.toPtt
 import net.mamoe.mirai.internal.message.toRichTextElems
 import net.mamoe.mirai.internal.network.Packet
 import net.mamoe.mirai.internal.network.QQAndroidClient
