+++ mirai-core/src/commonMain/resources/META-INF/services/net.mamoe.mirai.message.data.OfflineAudio.Factory
@@ -1,5 +1,5 @@
 #
-# Copyright 2019-2021 Mamoe Technologies and contributors.
+# Copyright 2019-2022 Mamoe Technologies and contributors.
 #
 # 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
 # Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
@@ -7,4 +7,4 @@
 # https://github.com/mamoe/mirai/blob/dev/LICENSE
 #
 
-net.mamoe.mirai.internal.message.OfflineAudioFactoryImpl
+net.mamoe.mirai.internal.message.data.OfflineAudioFactoryImpl
