+++ mirai-core/src/commonTest/kotlin/message/AudioTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2021 Mamoe Technologies and contributors.
+ * Copyright 2019-2022 Mamoe Technologies and contributors.
  *
  * 此源代码的使用受 GNU AFFERO GENERAL PUBLIC LICENSE version 3 许可证的约束, 可以在以下链接找到该许可证.
  * Use of this source code is governed by the GNU AGPLv3 license that can be found through the following link.
@@ -9,8 +9,10 @@
 
 package net.mamoe.mirai.internal.message
 
-import net.mamoe.mirai.internal.message.OnlineAudioImpl.Companion.DOWNLOAD_URL
-import net.mamoe.mirai.internal.message.OnlineAudioImpl.Companion.refineUrl
+import net.mamoe.mirai.internal.message.data.OfflineAudioImpl
+import net.mamoe.mirai.internal.message.data.OnlineAudioImpl
+import net.mamoe.mirai.internal.message.data.OnlineAudioImpl.Companion.DOWNLOAD_URL
+import net.mamoe.mirai.internal.message.data.OnlineAudioImpl.Companion.refineUrl
 import net.mamoe.mirai.internal.network.protocol.data.proto.ImMsgBody
 import net.mamoe.mirai.internal.test.AbstractTest
 import net.mamoe.mirai.message.data.AudioCodec
